Changed int to long

+++ src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
     public static void main(String[] args) {
 
-        int[] sales = new int[]{1, 5, 200, 400};
+        long[] sales = new long[]{1, 5, 200, 400};
         SalesManager saleManager = new SalesManager(sales);
-        int max = saleManager.max();
+        long max = saleManager.max();
         System.out.println(max);
     }
 }
